requirements.txt
absl-py==2.1.0
astunparse==1.6.3
bleach==6.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
flatbuffers==24.3.25
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.2
h5py==3.11.0
idna==3.7
joblib==1.4.0
kaggle==1.6.12
keras==3.3.3
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.8
numpy==1.26.4
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
protobuf==4.25.3
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
requests==2.31.0
rich==13.7.1
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
sklearn-preprocessing==0.1.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
text-unidecode==1.3
threadpoolctl==3.5.0
tqdm==4.66.2
typing_extensions==4.11.0
urllib3==2.2.1
webencodings==0.5.1
Werkzeug==3.0.2
wrapt==1.16.0
